change Injects method
