+++ src/background/background.ts
@@ -1,8 +1,5 @@
-import {ChromeMessage, Sender} from '../types';
+export {}
 
-/** Fired when the extension is first installed,
- *  when the extension is updated to a new version,
- *  and when Chrome is updated to a new version. */
 chrome.runtime.onInstalled.addListener((details) => {
     console.log('[background.js] onInstalled', details);
 });
@@ -15,46 +12,15 @@ chrome.runtime.onStartup.addListener(() => {
     console.log('[background.js] onStartup')
 });
 
-/**
- *  Sent to the event page just before it is unloaded.
- *  This gives the extension opportunity to do some clean up.
- *  Note that since the page is unloading,
- *  any asynchronous operations started while handling this event
- *  are not guaranteed to complete.
- *  If more activity for the event page occurs before it gets
- *  unloaded the onSuspendCanceled event will
- *  be sent and the page won't be unloaded. */
 chrome.runtime.onSuspend.addListener(() => {
     console.log('[background.js] onSuspend')
 });
 
-let tabId: number;
-
-chrome.webRequest.onCompleted.addListener(function (details) {
-    const parsedUrl = new URL(details.url);
-    if (tabId) {
-        const msg: ChromeMessage = {
-            from: Sender.Content,
-            message: parsedUrl.pathname,
-        }
-        chrome.tabs.sendMessage(tabId, msg)
+chrome.tabs.onUpdated.addListener(function (tabId, changeInfo, tab) {
+    if (tab.active) {
+        chrome.scripting.executeScript({
+            target: {tabId: tabId},
+            files: ['./static/js/myAnimeLink.js'],
+        });
     }
-}, {
-    urls: ['*://shikimori.one/animes/*', '*://shikimori.org/animes/*']
-});
-
-chrome.webNavigation.onHistoryStateUpdated.addListener(
-    details => {
-        tabId = details.tabId;
-    },
-    {
-        url: [
-            {
-                hostSuffix: 'shikimori.one'
-            },
-            {
-                hostSuffix: 'shikimori.org'
-            }
-        ]
-    }
-);
+})
